+++ src/scripts/main.js
@@ -2,6 +2,7 @@
 
 var loadLate = require('./loadLate');
 var subAndRun = require('./subAndRun');
+var mapboxQuery = require('./mapboxQuery');
 
 var subHub = subAndRun();
 
@@ -11,19 +12,7 @@ L.mapbox.accessToken = 'pk.eyJ1IjoiY2ZwYiIsImEiOiJodmtiSk5zIn0.VkCynzmVYcLBxbyHz
 var map = L.mapbox.map('map', 'mapbox.streets').setView([38, -122], 10);
 var features = L.mapbox.featureLayer(null).addTo(map);
 
-function randomBayPoint(){
-  var geo = {
-    "type": "Point",
-    "properties": {
-      "name": "Random Point"
-    }
-  };
-  var r1 = Math.random()/4;
-  var r2 = Math.random()/4;
 
-  geo.coordinates = [-122.12+r1, 37.87+r2];
-  return JSON.stringify(geo);
-}
 
 function flipCoords(arr){
   return [arr[1],arr[0]];
@@ -49,7 +38,7 @@ subHub.subscribe(function(result){
 var inp = document.getElementById('inp');
 inp.addEventListener('keydown',function(e){
   if(e.keyCode === 13){
-    var query = this.value;
+    var query = mapboxQuery(this.value);
     this.value = '';
     grass.geocode(query, process);
   }
